# openapi/analysis.yaml
...
openapi: 3.1.0
info:
  title: Unclaimed API
  version: 0.1.0-preview
  summary: Mounted safe-mode wallet analysis preview
  description: "Contract for the implemented POST /api/v1/check-wallet route.
    Analysis only: every opportunity has executionSupported: false and
    executionSession is null. No build, record, usage, payment or transaction
    execution capability is offered. This artifact does not establish deployment
    or live-provider readiness. The separate preview.yaml retains the future
    design."
servers:
  - url: /api/v1
    description: Relative to the operator-provided platform origin; no deployment is
      implied.
security:
  - apiKey: []
...
paths:
  /check-wallet:
    post:
      tags:
        - analysis
      operationId: checkWallet
      summary: Analyze one wallet page
      description: Analyze a frozen wallet page in safe mode. Required lookup failures
        return an incomplete error without a partial or invoiceable page. At
        most 400 raw token accounts; each page accepts limit 1–200 (default 20).
        First-page excess results may overlap cleanup; alternative groups must
        not be added together. Idempotency is mandatory, including retries.
        Execution is unavailable.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/SdkVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckWalletRequest"
      responses:
        "200":
          description: Complete page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckWalletResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: |
            Revoked key, or `mode_not_permitted`: the key lacks the `max`
            capability. Never silently downgraded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "410":
          description: "snapshot_expired: the shared 60-minute snapshot and cursor
            lifetime expired; start a new first page. No execution session is
            issued."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "413":
          $ref: "#/components/responses/TooLarge"
        "415":
          description: Unsupported JSON content encoding or charset. Journaled without
            admission.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: |
            `mode_not_available` (max analysis is not mounted yet),
            `invalid_wallet`, `cursor_mismatch`, `airdrops_not_available`
            (Preview), `wallet_too_large` (Preview: more than 400 raw token
            accounts), or a v1 module limit exceeded. Not invoiceable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          description: Sanitized internal analysis failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: |
            `incomplete` (Preview: a required lookup on this page failed after
            bounded retry; retry with the same cursor), `upstream_unavailable`,
            or `platform_unavailable` (retryable persistence/service failure).
            Retry platform failures with the same Idempotency-Key and request:
            an uncertain commit may already have a durable response to replay.
            The error response itself does not publish an invoiceable page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "504":
          $ref: "#/components/responses/Timeout"
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: Customer API key. Never sent to `/build` or `/executions/record`.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Required, 1–128 printable non-space ASCII characters. Same
        customer, key and canonical body replays the stored response without
        another admission or invoiceable outcome. A separate non-admitted,
        noninvoiceable replay usage row is recorded.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: ^[!-~]+$
    SdkVersion:
      name: X-Unclaimed-SDK-Version
      in: header
      required: false
      description: Diagnostics only; never changes results or metering.
      schema:
        type: string
        maxLength: 64
  schemas:
    CheckWalletRequest:
      type: object
      additionalProperties: false
      required:
        - wallet
      properties:
        wallet:
          $ref: "#/components/schemas/Base58"
        mode:
          type: string
          enum:
            - safe
          default: safe
          description: "Only safe analysis is available. Max requests are refused: 403
            without capability, 422 with capability. Never silently downgraded."
        includeAirdrops:
          type: boolean
          default: false
          description: "Preview: must be false; true returns 422 `airdrops_not_available`."
          const: false
        sellAndCloseMaxUsd:
          type: string
          default: "1.00"
          pattern: ^(?:[1-9](?:\.\d{1,6})?|10(?:\.0{1,6})?)$
          description: Accepted for v1 compatibility. Preview never returns an executable
            sell.
        limit:
          type: integer
          minimum: 1
          maximum: 200
          default: 20
        cursor:
          type:
            - string
            - "null"
          description: Opaque `nextCursor` from a prior page. Omit or null for the first
            page.
          pattern: ^ucur_[A-Za-z0-9_-]{43}$
    Base58:
      type: string
      pattern: ^[1-9A-HJ-NP-Za-km-z]{32,44}$
    CheckWalletResponse:
      type: object
      required:
        - requestId
        - apiVersion
        - rulesetVersion
        - assetIntelligenceRulesetVersion
        - analyzedAt
        - complete
        - data
        - billing
      properties:
        requestId:
          type: string
        apiVersion:
          type: string
          enum:
            - v1-preview
        rulesetVersion:
          type: string
        assetIntelligenceRulesetVersion:
          type: "null"
          description: "Preview: always null; no Asset Intelligence is served."
        analyzedAt:
          type: string
          format: date-time
        complete:
          type: boolean
          enum:
            - true
          description: Always true on a 200. Incomplete pages are refused, not served.
        data:
          type: object
          required:
            - wallet
            - mode
            - summary
            - pagination
            - items
            - airdrops
            - executionSession
            - limits
          properties:
            wallet:
              $ref: "#/components/schemas/Base58"
            mode:
              type: string
              enum:
                - safe
            summary:
              type: object
              description: Presentation-only, current page plus first-page wallet modules.
                Never a whole-wallet total while `hasMore` is true.
              required:
                - scope
                - analyzedAssetAccounts
                - opportunityCount
                - protectedAssetCount
                - totalsByAsset
                - estimatedNetValueUsd
              properties:
                scope:
                  type: string
                  enum:
                    - page
                analyzedAssetAccounts:
                  type: integer
                opportunityCount:
                  type: integer
                protectedAssetCount:
                  type: integer
                totalsByAsset:
                  type: array
                  items:
                    type: object
                    required:
                      - asset
                      - decimals
                      - grossBaseUnits
                      - serviceFeeBaseUnits
                      - netBaseUnits
                    properties:
                      asset:
                        type: string
                      mint:
                        $ref: "#/components/schemas/Base58"
                      decimals:
                        type: integer
                      grossBaseUnits:
                        $ref: "#/components/schemas/IntegerString"
                      serviceFeeBaseUnits:
                        $ref: "#/components/schemas/IntegerString"
                      netBaseUnits:
                        $ref: "#/components/schemas/IntegerString"
                estimatedNetValueUsd:
                  type:
                    - string
                    - "null"
                  description: Null whenever any required price input is missing or stale.
            pagination:
              type: object
              required:
                - returned
                - hasMore
                - nextCursor
                - walletModulesIncluded
                - expiresAt
              properties:
                returned:
                  type: integer
                  description: Raw token accounts analyzed on this page, including those with no
                    opportunity.
                hasMore:
                  type: boolean
                nextCursor:
                  type:
                    - string
                    - "null"
                walletModulesIncluded:
                  type: boolean
                  description: True only on the first page.
                expiresAt:
                  type: string
                  format: date-time
                  description: Shared snapshot and cursor expiry. No execution session is issued.
            items:
              type: array
              items:
                $ref: "#/components/schemas/WalletItem"
            airdrops:
              type: "null"
            executionSession:
              type: "null"
              description: Always null in the current analysis-only increment; sessions and
                builders are outside this stage.
            limits:
              type: object
              required:
                - maxRawTokenAccounts
                - pageLimitMax
              properties:
                maxRawTokenAccounts:
                  type: integer
                  enum:
                    - 400
                  description: Preview cap, enforced before enrichment.
                pageLimitMax:
                  type: integer
                  enum:
                    - 200
        billing:
          $ref: "#/components/schemas/Billing"
    IntegerString:
      type: string
      pattern: ^[0-9]+$
      description: Integer base units as a decimal string. Never a JSON number.
    WalletItem:
      type: object
      description: One discriminated item. Preview omits `discardEvidence` and
        `classification.reputation` entirely rather than nulling them.
      required:
        - id
        - kind
        - source
        - classification
        - opportunity
      properties:
        id:
          type: string
        kind:
          type: string
          enum:
            - token_account
            - excess_lamports
        source:
          type: string
          enum:
            - spl_token
            - token_2022
            - excess_lamports
          description: For disappeared or unsupported accounts, the platform retains the
            source from its address-bound frozen inventory. It never supplies
            missing current account facts from that snapshot.
        targetAccount:
          $ref: "#/components/schemas/Base58"
          description: Exact target of an excess-lamports item. No current token balance
            is inferred from frozen inventory.
        alternativeGroupId:
          type: string
          description: Present on every excess-lamports item for a token account and on
            the token-account item for the same account. The two are
            alternatives, never additive.
        asset:
          type: object
          required:
            - mint
            - assetType
            - decimals
          properties:
            mint:
              $ref: "#/components/schemas/Base58"
            assetType:
              type: string
              enum:
                - fungible
                - token_account_collectible
                - unknown
            symbol:
              type:
                - string
                - "null"
            decimals:
              type: integer
        account:
          type: object
          required:
            - address
            - program
            - balanceBaseUnits
            - rentLamports
          additionalProperties: false
          properties:
            address:
              $ref: "#/components/schemas/Base58"
            program:
              type: string
              enum:
                - spl_token
                - token_2022
            balanceBaseUnits:
              $ref: "#/components/schemas/IntegerString"
            rentLamports:
              $ref: "#/components/schemas/IntegerString"
        classification:
          type: object
          required:
            - disposition
            - confidence
            - reviewRequired
            - reasons
            - protections
          properties:
            disposition:
              type: string
              enum:
                - closeable_empty
                - withdrawable
                - protected
                - review_only
                - no_op
              description: Safe analysis only. Max-mode burn candidates are unavailable. Sell
                candidates are returned as `review_only`.
            confidence:
              type: string
              enum:
                - known
                - high
                - low
            reviewRequired:
              type: boolean
            reasons:
              type: array
              items:
                type: string
                enum:
                  - zero_balance
                  - priced_dust_within_policy
                  - sell_not_supported_in_preview
                  - above_dust_threshold
                  - above_net_rent
                  - frozen
                  - collectible
                  - lp_or_position_shaped
                  - receipt_token_policy
                  - dangerous_token_2022_extension
                  - account_missing
                  - unsupported_account
                  - positive_wsol
                  - unpriced
                  - completed_no_sell_route
                  - unsupported_signer
                  - non_positive_net
                  - excess_above_rent_exempt_minimum
            protections:
              type: array
              items:
                type: string
                enum:
                  - frozen
                  - collectible
                  - lp_position
                  - receipt_token
                  - token_2022_extension
                  - unpriced
                  - account_state
        market:
          type: object
          description: Present on fungible token-account items.
          required:
            - quoteStatus
            - sellRouteAvailable
            - executableQuoteUsd
            - quotedAt
          properties:
            quoteStatus:
              type: string
              enum:
                - complete
                - unavailable
            sellRouteAvailable:
              type:
                - boolean
                - "null"
              description: Null when a route was not requested; false only for a completed
                no-route lookup.
            executableQuoteUsd:
              oneOf:
                - type: "null"
                - $ref: "#/components/schemas/DecimalUsd"
            quotedAt:
              type:
                - string
                - "null"
              format: date-time
          allOf:
            - if:
                properties:
                  quoteStatus:
                    const: unavailable
              then:
                properties:
                  sellRouteAvailable:
                    const: null
                  executableQuoteUsd:
                    const: null
                  quotedAt:
                    const: null
              else:
                properties:
                  sellRouteAvailable:
                    type: boolean
                  quotedAt:
                    type: string
                    format: date-time
            - if:
                properties:
                  sellRouteAvailable:
                    const: false
              then:
                properties:
                  executableQuoteUsd:
                    const: null
        opportunity:
          oneOf:
            - $ref: "#/components/schemas/BurnAndCloseOpportunity"
            - $ref: "#/components/schemas/RecoverExcessLamportsOpportunity"
            - $ref: "#/components/schemas/NoActionOpportunity"
          discriminator:
            propertyName: action
            mapping:
              burn_and_close: "#/components/schemas/BurnAndCloseOpportunity"
              recover_excess_lamports: "#/components/schemas/RecoverExcessLamportsOpportunity"
              none: "#/components/schemas/NoActionOpportunity"
      allOf:
        - if:
            properties:
              kind:
                const: excess_lamports
          then:
            required:
              - targetAccount
              - alternativeGroupId
            not:
              required:
                - account
            properties:
              source:
                const: excess_lamports
              classification:
                properties:
                  disposition:
                    const: withdrawable
                  confidence:
                    const: known
                  reviewRequired:
                    const: true
                  reasons:
                    const:
                      - excess_above_rent_exempt_minimum
                  protections:
                    maxItems: 0
              opportunity:
                properties:
                  action:
                    const: recover_excess_lamports
          else:
            not:
              required:
                - targetAccount
            properties:
              source:
                enum:
                  - spl_token
                  - token_2022
              classification:
                properties:
                  disposition:
                    enum:
                      - closeable_empty
                      - protected
                      - review_only
                      - no_op
                  reasons:
                    not:
                      contains:
                        const: excess_above_rent_exempt_minimum
              opportunity:
                properties:
                  action:
                    enum:
                      - none
                      - burn_and_close
        - if:
            required:
              - account
            properties:
              account:
                properties:
                  program:
                    const: spl_token
          then:
            properties:
              source:
                const: spl_token
        - if:
            required:
              - account
            properties:
              account:
                properties:
                  program:
                    const: token_2022
          then:
            properties:
              source:
                const: token_2022
        - if:
            required:
              - market
          then:
            required:
              - account
              - asset
            properties:
              kind:
                const: token_account
              asset:
                properties:
                  assetType:
                    const: fungible
        - if:
            properties:
              opportunity:
                properties:
                  action:
                    const: burn_and_close
          then:
            required:
              - account
              - asset
            properties:
              kind:
                const: token_account
              source:
                enum:
                  - spl_token
                  - token_2022
              classification:
                properties:
                  disposition:
                    enum:
                      - closeable_empty
                  reviewRequired:
                    const: true
                  protections:
                    maxItems: 0
        - if:
            properties:
              classification:
                properties:
                  disposition:
                    const: closeable_empty
          then:
            required:
              - account
              - asset
            properties:
              account:
                properties:
                  balanceBaseUnits:
                    const: "0"
              opportunity:
                properties:
                  action:
                    const: burn_and_close
                  reviewedBalanceBaseUnits:
                    const: "0"
        - if:
            properties:
              classification:
                properties:
                  reasons:
                    contains:
                      enum:
                        - account_missing
                        - unsupported_account
          then:
            properties:
              classification:
                properties:
                  disposition:
                    const: protected
                  protections:
                    contains:
                      const: account_state
              opportunity:
                properties:
                  action:
                    const: none
            not:
              required:
                - market
    DecimalUsd:
      type: string
      pattern: ^[0-9]+\.[0-9]{6}$
      description: USD with six decimals, as a string.
...
    RecoverExcessLamportsOpportunity:
      type: object
      required:
        - action
        - destructive
        - executionSupported
        - valueComponents
        - costs
        - estimatedNetValueUsd
        - estimated
      properties:
        action:
          type: string
          enum:
            - recover_excess_lamports
        destructive:
          type: boolean
          enum:
            - false
        executionSupported:
          type: boolean
          const: false
          description: Execution is unavailable in this analysis-only stage.
        valueComponents:
          type: array
          minItems: 1
          items:
            allOf:
              - $ref: "#/components/schemas/ValueComponent"
              - properties:
                  source:
                    const: excess_lamports
        costs:
          type: "null"
          description: Null while the analysis adapter supplies no network/funding
            estimate. Never infer zero cost.
        estimatedNetValueUsd:
          type: "null"
          description: No USD valuation is supplied by this analysis adapter.
        estimated:
          type: boolean
          enum:
            - true
...
    Billing:
      type: object
      description: |
        Preview: `mode` is always `metered_preview`. `chargedUsd` is the list
        price recorded for this complete response and is invoiceable by manual
        invoice only. `balanceRemainingUsd` is always null because no prepaid
        balance exists in the preview.
      required:
        - mode
        - currency
        - billableUnits
        - baseChargeUsd
        - addOnChargeUsd
        - chargedUsd
        - balanceRemainingUsd
      properties:
        mode:
          type: string
          enum:
            - metered_preview
        currency:
          type: string
          enum:
            - USD
        billableUnits:
          type: integer
          minimum: 1
        baseChargeUsd:
          $ref: "#/components/schemas/DecimalUsd"
        addOnChargeUsd:
          $ref: "#/components/schemas/DecimalUsd"
        chargedUsd:
          $ref: "#/components/schemas/DecimalUsd"
        balanceRemainingUsd:
          type: "null"
    ErrorEnvelope:
      type: object
      required:
        - requestId
        - error
      properties:
        requestId:
          type: string
        error:
          type: object
          required:
            - code
            - message
            - retryable
          properties:
            code:
              type: string
              enum:
                - unauthorized
                - key_revoked
                - mode_not_permitted
                - mode_not_available
                - idempotency_key_reused
                - request_in_progress
                - snapshot_expired
                - request_too_large
                - invalid_wallet
                - cursor_mismatch
                - airdrops_not_available
                - wallet_too_large
                - module_limit_exceeded
                - rate_limited
                - internal_error
                - invalid_idempotency_key
                - platform_unavailable
                - invalid_request
                - incomplete
                - upstream_unavailable
                - deadline_exceeded
            message:
              type: string
            retryable:
              type: boolean
            details:
              type: object
              additionalProperties: true
              description: Never contains upstream URLs, credentials, raw RPC messages, or
                provider names.
  responses:
    BadRequest:
      description: Malformed JSON or unsupported request shape.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    IdempotencyConflict:
      description: >
        `idempotency_key_reused` (not retryable: same key, different endpoint or
        body) or

        `request_in_progress` (retryable: the owning request is still running;
        keep the same key and retry after `Retry-After`).
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    TooLarge:
      description: Request body too large.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: Customer rate or concurrency limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Timeout:
      description: 30-second deadline exceeded. Not invoiceable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"